requirements.txt
adal==1.2.7
aiohttp==3.8.6
aiosignal==1.3.1
alembic==1.12.0
altair==5.1.2
annotated-types==0.6.0
anyio==4.0.0
apache-airflow==2.7.2
apache-airflow-providers-common-sql==1.7.2
apache-airflow-providers-ftp==3.5.2
apache-airflow-providers-http==4.5.2
apache-airflow-providers-imap==3.3.2
apache-airflow-providers-microsoft-azure==8.0.0
apache-airflow-providers-sqlite==3.4.3
apispec==6.3.0
argcomplete==3.1.2
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
azure-batch==14.0.0
azure-common==1.1.28
azure-core==1.29.4
azure-cosmos==4.5.1
azure-datalake-store==0.0.53
azure-identity==1.14.1
azure-keyvault-secrets==4.7.0
azure-kusto-data==4.2.0
azure-kusto-ingest==4.2.0
azure-mgmt-containerinstance==10.1.0
azure-mgmt-containerregistry==10.2.0
azure-mgmt-core==1.4.0
azure-mgmt-cosmosdb==9.3.0
azure-mgmt-datafactory==3.1.0
azure-mgmt-datalake-nspkg==3.0.1
azure-mgmt-datalake-store==0.5.0
azure-mgmt-kusto==3.2.0
azure-mgmt-nspkg==3.0.2
azure-mgmt-resource==23.0.1
azure-mgmt-storage==21.1.0
azure-nspkg==3.0.2
azure-servicebus==7.11.3
azure-storage-blob==12.18.3
azure-storage-file-datalake==12.13.2
azure-storage-file-share==12.14.2
azure-storage-queue==12.7.3
azure-synapse-spark==0.7.0
Babel==2.13.0
backoff==1.10.0
blinker==1.6.3
cachelib==0.9.0
cachetools==5.3.1
cattrs==23.1.2
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
clickclick==20.10.2
colorama==0.4.6
colorlog==4.8.0
ConfigUpdater==3.1.1
connexion==2.14.2
cron-descriptor==1.4.0
croniter==1.4.1
cryptography==41.0.4
Deprecated==1.2.14
dill==0.3.1.1
dnspython==2.4.2
docutils==0.20.1
email-validator==1.3.1
exceptiongroup==1.1.3
Flask==2.2.5
Flask-AppBuilder==4.3.6
Flask-Babel==2.0.0
Flask-Caching==2.0.2
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.5.0
Flask-Login==0.6.2
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.1
frozenlist==1.4.0
gitdb==4.0.11
GitPython==3.1.40
google-re2==1.1
googleapis-common-protos==1.60.0
graphviz==0.20.1
greenlet==3.0.0
grpcio==1.59.0
gunicorn==21.2.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
ijson==3.2.3
importlib-metadata==6.8.0
importlib-resources==6.1.0
inflection==0.5.1
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
lazy-object-proxy==1.9.0
limits==3.6.0
linkify-it-py==2.0.2
lockfile==0.12.2
Mako==1.2.4
Markdown==3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-oneofschema==3.0.1
marshmallow-sqlalchemy==0.26.1
mdit-py-plugins==0.4.0
mdurl==0.1.2
msal==1.24.1
msal-extensions==1.0.0
msrest==0.7.1
msrestazure==0.6.4
multidict==6.0.4
numpy==1.26.1
oauthlib==3.2.2
opentelemetry-api==1.20.0
opentelemetry-exporter-otlp==1.20.0
opentelemetry-exporter-otlp-proto-common==1.20.0
opentelemetry-exporter-otlp-proto-grpc==1.20.0
opentelemetry-exporter-otlp-proto-http==1.20.0
opentelemetry-proto==1.20.0
opentelemetry-sdk==1.20.0
opentelemetry-semantic-conventions==0.41b0
ordered-set==4.1.0
packaging==23.2
pandas==2.1.1
pathspec==0.11.2
pendulum==2.1.2
Pillow==10.1.0
pluggy==1.3.0
portalocker==2.8.2
prison==0.2.1
protobuf==4.21.12
psutil==5.9.5
pyarrow==13.0.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
PyJWT==2.8.0
python-daemon==3.0.1
python-dateutil==2.8.2
python-nvd3==0.15.0
python-slugify==8.0.1
pytz==2023.3.post1
pytzdata==2020.1
PyYAML==6.0.1
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rich==13.6.0
rich-argparse==1.3.0
rpds-py==0.10.4
scikit-learn==1.3.1
scipy==1.11.3
setproctitle==1.3.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
SQLAlchemy==1.4.49
SQLAlchemy-JSONField==1.0.1.post0
SQLAlchemy-Utils==0.41.1
sqlparse==0.4.4
streamlit==1.27.2
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.3.0
text-unidecode==1.3
threadpoolctl==3.2.0
toml==0.10.2
toolz==0.12.0
tornado==6.3.3
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
uc-micro-py==1.0.2
unicodecsv==0.14.1
urllib3==1.26.17
validators==0.22.0
watchdog==3.0.0
Werkzeug==2.2.3
wrapt==1.15.0
WTForms==3.0.1
yarl==1.9.2
zipp==3.17.0
